third_party/WebKit/LayoutTests/media/track/track-cue-rendering-on-resize.html

<!DOCTYPE html>
<title>Test that line height of cue element isn't overridden by author CSS.</title>
<script src="../media-controls.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
body { line-height: 108px; }
</style>
<video>
    <track src="captions-webvtt/captions-snap-to-lines-set.vtt" kind="captions" default>
</video>
<script>
async_test(function(t) {
    var video = document.querySelector('video');
    video.src = '../content/test.ogv';

    video.onseeked = t.step_func_done(function() {
        var multiLineCueDisplayElement = textTrackDisplayElement(video);
        assert_equals(getComputedStyle(multiLineCueDisplayElement).lineHeight, 'normal');
    });

    video.currentTime = 1.25;
});
</script>
